requirements.txt
black==23.11.0
celery==5.3.6
certifi==2023.11.17
Django==4.2.7
django-celery-beat==2.5.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
isort==5.12.0
Pillow==10.1.0
psycopg2-binary==2.9.9
requests==2.31.0


